CustomFitProject/accounts/views.py
# customFit 앱 구현 테스트용 임시 모델 입니다!

from rest_framework import viewsets
from .models import User
from .serializers import UserSerializer

class UserViewSet(viewsets.ModelViewSet):
    queryset = User.objects.all()
    serializer_class = UserSerializer
